+++ packages/file_selector/file_selector_platform_interface/test/file_selector_platform_interface_test.dart
@@ -31,14 +31,16 @@ void main() {
   });
 
   group('getDirectoryPathWithOptions', () {
-    test('Should throw unimplemented exception', () async {
-      final FileSelectorPlatform fileSelector = ExtendsFileSelectorPlatform();
+    test('Should fall back to getDirectoryPath by default', () async {
+      final FileSelectorPlatform fileSelector =
+          OldFileSelectorPlatformImplementation();
 
-      await expectLater(() async {
-        return fileSelector.getDirectoryPathWithOptions(
-          const FileDialogOptions(),
-        );
-      }, throwsA(isA<UnimplementedError>()));
+      final String? result = await fileSelector.getDirectoryPathWithOptions(
+        const FileDialogOptions(),
+      );
+
+      // Should call the old method and return its result
+      expect(result, OldFileSelectorPlatformImplementation.directoryPath);
     });
   });
 
@@ -53,14 +55,17 @@ void main() {
   });
 
   group('getDirectoryPathsWithOptions', () {
-    test('Should throw unimplemented exception', () async {
-      final FileSelectorPlatform fileSelector = ExtendsFileSelectorPlatform();
+    test('Should fall back to getDirectoryPaths by default', () async {
+      final FileSelectorPlatform fileSelector =
+          OldFileSelectorPlatformImplementation();
 
-      await expectLater(() async {
-        return fileSelector.getDirectoryPathsWithOptions(
-          const FileDialogOptions(),
-        );
-      }, throwsA(isA<UnimplementedError>()));
+      final List<String> result = await fileSelector
+          .getDirectoryPathsWithOptions(const FileDialogOptions());
+
+      // Should call the old method and return its result
+      expect(result, <String>[
+        OldFileSelectorPlatformImplementation.directoryPath,
+      ]);
     });
   });
 
@@ -79,7 +84,8 @@ class ExtendsFileSelectorPlatform extends FileSelectorPlatform {}
 
 class OldFileSelectorPlatformImplementation extends FileSelectorPlatform {
   static const String savePath = '/a/path';
-  // Only implement the deprecated getSavePath.
+  static const String directoryPath = '/a/directory';
+
   @override
   Future<String?> getSavePath({
     List<XTypeGroup>? acceptedTypeGroups,
@@ -88,5 +94,21 @@ class OldFileSelectorPlatformImplementation extends FileSelectorPlatform {
     String? confirmButtonText,
   }) async {
     return savePath;
+  }
+
+  @override
+  Future<String?> getDirectoryPath({
+    String? initialDirectory,
+    String? confirmButtonText,
+  }) async {
+    return directoryPath;
+  }
+
+  @override
+  Future<List<String>> getDirectoryPaths({
+    String? initialDirectory,
+    String? confirmButtonText,
+  }) async {
+    return <String>[directoryPath];
   }
 }