+++ packages/file_selector/file_selector/CHANGELOG.md
@@ -1,11 +1,10 @@
 ## 1.1.0
 
-* Adds `canCreateDirectories` parameter to `FileDialogOptions` to control the visibility of the New Folder button in file dialogs on supported platforms.
+* Adds `canCreateDirectories` param to `getDirectoryPath` and `getDirectoryPaths` to control the visibility of the New Folder button in file dialogs on supported platforms.
 
 ## 1.0.4
 
 * Updates the example app and README examples to work on iOS.
-
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 * Updates README to indicate that Andoid SDK <21 is no longer supported.